refactor: 카드 생성시 등록과 취소 버튼 함수 NewCardForm에서 정의
